boot-basic/autoconf/src/main/java/org/spring/boot/HelloAutoConfiguration.java
	package org.spring.boot;

import org.spring.helloservice.ConsoleHelloService;
import org.spring.helloservice.HelloService;
import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
import org.springframework.boot.context.properties.EnableConfigurationProperties;
import org.springframework.context.annotation.Bean;
import org.springframework.context.annotation.Configuration;

@Configuration
@ConditionalOnClass(HelloService.class)
@EnableConfigurationProperties(HelloProperties.class)
public class HelloAutoConfiguration {

	@Bean
	@ConditionalOnMissingBean	
	public HelloService helloService(HelloProperties properties) {
		return new ConsoleHelloService(
				properties.getPrefix(),
				properties.getSuffix()
				);
	}

}
